Rename eggUnk.cpp to eggThreadMonitor.cpp
