+++ include/egg/core/eggThreadMonitor.h
